Mini Project
Selecting the items on the menu, they appear together with their prices
in the separate div block in random colors. The total price of all the
dishes appears in the alert window after pressing “Add money!” button.
Also you can see a location of this restaurant slicking on its name.